Все эмбеды переведены на новый стандарт, добавлена проверка на аргумент в say

+++ src/main/java/commands/SayCommand.java
@@ -1,5 +1,6 @@
 package commands;
 
+import api.BasicEmbed;
 import api.models.command.Command;
 import api.models.command.DiscordCommand;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
@@ -9,8 +10,14 @@ public class SayCommand implements Command {
 
     @Override
     public void doCommand(MessageReceivedEvent msg_event, String[] arguments){
-        msg_event.getMessage().delete().queue();
-        msg_event.getChannel().sendMessage(String.join(" ", arguments)).queue();
+        if (arguments.length != 0) {
+            msg_event.getMessage().delete().queue();
+            msg_event.getChannel().sendMessage(String.join(" ", arguments)).queue();
+        } else {
+            BasicEmbed errorEmbed = new BasicEmbed("error");
+            errorEmbed.setDescription("О, а можете меня научить ничего не говорить по команде?");
+            msg_event.getChannel().sendMessage(errorEmbed.build()).queue();
+        }
     }
 
 }
